requirements.txt
fastapi
uvicorn
pydantic
pysmb
python-dotenv